backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.9
httpx>=0.27.0
pydantic[email]>=2.9.0
slowapi>=0.1.9
